+++ amethyst_rendy/src/pass/flat2d.rs
@@ -109,7 +109,7 @@ impl<B: Backend> RenderGroup<B, Resources> for DrawFlat2D<B> {
         let (
             sprite_sheet_storage,
             tex_storage,
-            visibilities,
+            visibility,
             hiddens,
             hidden_props,
             sprite_renders,
@@ -118,7 +118,7 @@ impl<B: Backend> RenderGroup<B, Resources> for DrawFlat2D<B> {
         ) = <(
             Read<'_, AssetStorage<SpriteSheet>>,
             Read<'_, AssetStorage<Texture>>,
-            Option<Read<'_, SpriteVisibility>>,
+            ReadExpect<'_, SpriteVisibility>,
             ReadStorage<'_, Hidden>,
             ReadStorage<'_, HiddenPropagate>,
             ReadStorage<'_, SpriteRender>,
@@ -133,70 +133,36 @@ impl<B: Backend> RenderGroup<B, Resources> for DrawFlat2D<B> {
 
         sprites_ref.clear_inner();
 
-        match visibilities {
-            None => {
-                #[cfg(feature = "profiler")]
-                profile_scope!("gather_novisibility");
+        {
+            #[cfg(feature = "profiler")]
+            profile_scope!("gather_visibility");
 
-                (
-                    &sprite_renders,
-                    &transforms,
-                    tints.maybe(),
-                    !&hiddens,
-                    !&hidden_props,
-                )
-                    .join()
-                    .filter_map(|(sprite_render, global, tint, _, _)| {
-                        let (batch_data, texture) = SpriteArgs::from_data(
-                            &tex_storage,
-                            &sprite_sheet_storage,
-                            &sprite_render,
-                            &global,
-                            tint,
-                        )?;
-                        let (tex_id, _) = textures_ref.insert(
-                            factory,
-                            resources,
-                            texture,
-                            hal::image::Layout::ShaderReadOnlyOptimal,
-                        )?;
-                        Some((tex_id, batch_data))
-                    })
-                    .for_each_group(|tex_id, batch_data| {
-                        sprites_ref.insert(tex_id, batch_data.drain(..))
-                    });
-            }
-            Some(ref visibility) => {
-                #[cfg(feature = "profiler")]
-                profile_scope!("gather_visibility");
-
-                (
-                    &sprite_renders,
-                    &transforms,
-                    &visibility.visible_unordered,
-                    tints.maybe(),
-                )
-                    .join()
-                    .filter_map(|(sprite_render, global, _, tint)| {
-                        let (batch_data, texture) = SpriteArgs::from_data(
-                            &tex_storage,
-                            &sprite_sheet_storage,
-                            &sprite_render,
-                            &global,
-                            tint,
-                        )?;
-                        let (tex_id, _) = textures_ref.insert(
-                            factory,
-                            resources,
-                            texture,
-                            hal::image::Layout::ShaderReadOnlyOptimal,
-                        )?;
-                        Some((tex_id, batch_data))
-                    })
-                    .for_each_group(|tex_id, batch_data| {
-                        sprites_ref.insert(tex_id, batch_data.drain(..))
-                    });
-            }
+            (
+                &sprite_renders,
+                &transforms,
+                tints.maybe(),
+                &visibility.visible_unordered,
+            )
+                .join()
+                .filter_map(|(sprite_render, global, tint, _)| {
+                    let (batch_data, texture) = SpriteArgs::from_data(
+                        &tex_storage,
+                        &sprite_sheet_storage,
+                        &sprite_render,
+                        &global,
+                        tint,
+                    )?;
+                    let (tex_id, _) = textures_ref.insert(
+                        factory,
+                        resources,
+                        texture,
+                        hal::image::Layout::ShaderReadOnlyOptimal,
+                    )?;
+                    Some((tex_id, batch_data))
+                })
+                .for_each_group(|tex_id, batch_data| {
+                    sprites_ref.insert(tex_id, batch_data.drain(..))
+                });
         }
 
         self.textures.maintain(factory, resources);