+++ amethyst_test/src/amethyst_application.rs
@@ -7,7 +7,6 @@ use amethyst::{
     error::Error,
     input::{BindingTypes, InputBundle},
     prelude::*,
-    renderer::types::DefaultBackend,
     shred::Resource,
     ui::UiBundle,
     utils::application_root_dir,
@@ -349,7 +348,7 @@ where
     /// * `B`: Type representing the input binding types.
     pub fn with_ui_bundles<B: BindingTypes>(self) -> Self {
         self.with_bundle(InputBundle::<B>::new())
-            .with_bundle(UiBundle::<DefaultBackend, B>::new())
+            .with_bundle(UiBundle::<B>::new())
     }
 
     /// Adds a resource to the `World`.