+++ examples/arc_ball_camera/main.rs
@@ -7,30 +7,19 @@ use amethyst::{
         shrev::{EventChannel, ReaderId},
         transform::{Transform, TransformBundle},
     },
-    ecs::prelude::{
-        Join, Read, ReadExpect, ReadStorage, Resources, System, SystemData, WriteStorage,
-    },
+    ecs::prelude::{Join, Read, ReadStorage, Resources, System, SystemData, WriteStorage},
     input::{
         is_key_down, InputBundle, InputEvent, ScrollDirection, StringBindings, VirtualKeyCode,
     },
     prelude::*,
     renderer::{
         palette::Srgb,
-        pass::{DrawShadedDesc, DrawSkyboxDesc},
-        rendy::{
-            factory::Factory,
-            graph::{
-                render::{RenderGroupDesc, SubpassBuilder},
-                GraphBuilder,
-            },
-            hal::{format::Format, image},
-            mesh::{Normal, Position, TexCoord},
-        },
+        plugins::{RenderShaded3D, RenderSkybox, RenderToWindow},
+        rendy::mesh::{Normal, Position, TexCoord},
         types::DefaultBackend,
-        GraphCreator, RenderingSystem,
+        RenderingBundle,
     },
     utils::{application_root_dir, scene::BasicScenePrefab},
-    window::{ScreenDimensions, Window, WindowBundle},
     Error,
 };
 
@@ -64,7 +53,7 @@ impl SimpleState for ExampleState {
 }
 
 struct CameraDistanceSystem {
-    event_reader: Option<ReaderId<InputEvent<String>>>,
+    event_reader: Option<ReaderId<InputEvent<StringBindings>>>,
 }
 
 impl CameraDistanceSystem {
@@ -75,7 +64,7 @@ impl CameraDistanceSystem {
 
 impl<'a> System<'a> for CameraDistanceSystem {
     type SystemData = (
-        Read<'a, EventChannel<InputEvent<String>>>,
+        Read<'a, EventChannel<InputEvent<StringBindings>>>,
         ReadStorage<'a, Transform>,
         WriteStorage<'a, ArcBallControlTag>,
     );
@@ -104,7 +93,7 @@ impl<'a> System<'a> for CameraDistanceSystem {
         Self::SystemData::setup(res);
 
         self.event_reader = Some(
-            res.fetch_mut::<EventChannel<InputEvent<String>>>()
+            res.fetch_mut::<EventChannel<InputEvent<StringBindings>>>()
                 .register_reader(),
         );
     }
@@ -115,14 +104,12 @@ fn main() -> Result<(), Error> {
 
     let app_root = application_root_dir()?;
 
-    let resources_directory = app_root.join("examples/assets");
-    let display_config_path =
-        app_root.join("examples/arc_ball_camera/resources/display_config.ron");
+    let assets_directory = app_root.join("examples/assets");
+    let display_config_path = app_root.join("examples/arc_ball_camera/config/display.ron");
 
-    let key_bindings_path = app_root.join("examples/arc_ball_camera/resources/input.ron");
+    let key_bindings_path = app_root.join("examples/arc_ball_camera/config/input.ron");
 
     let game_data = GameDataBuilder::default()
-        .with_bundle(WindowBundle::from_config_path(display_config_path))?
         .with(PrefabLoaderSystem::<MyPrefabData>::default(), "", &[])
         .with_bundle(TransformBundle::new().with_dep(&[]))?
         .with_bundle(
@@ -134,90 +121,17 @@ fn main() -> Result<(), Error> {
             "camera_distance_system",
             &["input_system"],
         )
-        .with_thread_local(RenderingSystem::<DefaultBackend, _>::new(
-            ExampleGraph::default(),
-        ));
-    let mut game = Application::build(resources_directory, ExampleState)?.build(game_data)?;
+        .with_bundle(
+            RenderingBundle::<DefaultBackend>::new()
+                .with_plugin(RenderToWindow::from_config_path(display_config_path))
+                .with_plugin(RenderShaded3D::default())
+                .with_plugin(RenderSkybox::with_colors(
+                    Srgb::new(0.82, 0.51, 0.50),
+                    Srgb::new(0.18, 0.11, 0.85),
+                )),
+        )?;
+
+    let mut game = Application::build(assets_directory, ExampleState)?.build(game_data)?;
     game.run();
     Ok(())
-}
-
-#[derive(Default)]
-struct ExampleGraph {
-    dimensions: Option<ScreenDimensions>,
-    surface_format: Option<Format>,
-    dirty: bool,
-}
-
-#[allow(clippy::map_clone)]
-impl GraphCreator<DefaultBackend> for ExampleGraph {
-    fn rebuild(&mut self, res: &Resources) -> bool {
-        // Rebuild when dimensions change, but wait until at least two frames have the same.
-        let new_dimensions = res.try_fetch::<ScreenDimensions>();
-        use std::ops::Deref;
-        if self.dimensions.as_ref() != new_dimensions.as_ref().map(|d| d.deref()) {
-            self.dirty = true;
-            self.dimensions = new_dimensions.map(|d| d.clone());
-            return false;
-        }
-        self.dirty
-    }
-
-    fn builder(
-        &mut self,
-        factory: &mut Factory<DefaultBackend>,
-        res: &Resources,
-    ) -> GraphBuilder<DefaultBackend, Resources> {
-        use amethyst::renderer::rendy::{
-            graph::present::PresentNode,
-            hal::command::{ClearDepthStencil, ClearValue},
-        };
-
-        self.dirty = false;
-
-        let window = <ReadExpect<'_, Window>>::fetch(res);
-        let surface = factory.create_surface(&window);
-        // cache surface format to speed things up
-        let surface_format = *self
-            .surface_format
-            .get_or_insert_with(|| factory.get_surface_format(&surface));
-        let dimensions = self.dimensions.as_ref().unwrap();
-        let window_kind =
-            image::Kind::D2(dimensions.width() as u32, dimensions.height() as u32, 1, 1);
-
-        let mut graph_builder = GraphBuilder::new();
-        let color = graph_builder.create_image(
-            window_kind,
-            1,
-            surface_format,
-            Some(ClearValue::Color([0.34, 0.36, 0.52, 1.0].into())),
-        );
-
-        let depth = graph_builder.create_image(
-            window_kind,
-            1,
-            Format::D32Sfloat,
-            Some(ClearValue::DepthStencil(ClearDepthStencil(1.0, 0))),
-        );
-
-        let opaque = graph_builder.add_node(
-            SubpassBuilder::new()
-                .with_group(DrawShadedDesc::new().builder())
-                .with_group(
-                    DrawSkyboxDesc::with_colors(
-                        Srgb::new(0.82, 0.51, 0.50),
-                        Srgb::new(0.18, 0.11, 0.85),
-                    )
-                    .builder(),
-                )
-                .with_color(color)
-                .with_depth_stencil(depth)
-                .into_pass(),
-        );
-
-        let _present = graph_builder
-            .add_node(PresentNode::builder(factory, surface, color).with_dependency(opaque));
-
-        graph_builder
-    }
 }