+++ docs/CHANGELOG.md
@@ -1,4 +1,5 @@
 # Change Log
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog][kc], and this project adheres to
@@ -10,7 +11,9 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 ## [Unreleased]
 
 ## Breaking changes
+
 * `Float` newtype removed, moved back to `f32` primitive for all values ([#1747])
+* `TextureProcessor` and `MeshProcessor` systems are now separated from `RenderingSystem` ([#1772])
 
 ### Added
 
@@ -19,9 +22,12 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 * Add `add_rectangle`, `add_rotated_rectangle`, `add_box`, `add_rotated_box`, `add_circle`, `add_rotated_circle`,
 `add_cylinder`, `add_rotated_cylinder` and `add_sphere` functions to `DebugLinesComponent`
 and the corresponding draw functions to `DebugLines`, to draw simple shapes with debug lines. ([#1766])
+* `InputEvent::AxisMoved` is sent upon button press / release. ([#1512], [#1797])
+* Added `RenderingBundle` with a rendering plugin system, making rendering setup easier ([#1772])
 
 ### Changed
 
+* Splitted the `/resources` directory of amethyst projects into `/assets` and `/config`. ([#1806])
 * Rename FPSCounter, FPSCounterBundle, FPSCounterSystem to FpsCounter, FpsCounterBundle, FpsCounterSystem. ([#1719])
 * Add Tint component support for sprites. ([#1756])
 * Remove remaining <N: RealField> type parameter on GameDataBuilder, add Debug derive to LoggerConfig ([#1758])
@@ -32,12 +38,13 @@ and the corresponding draw functions to `DebugLines`, to draw simple shapes with
 * Updated fluent to version 0.6. ([#1800])
 
 ### Fixed
+
 * Fix stack overflow on serializing `Box<dyn Format<_>>`. ([#1720])
-
-### Fixed
-
+ `InputEvent<T>` now takes in the `BindingTypes` as a type parameter. ([#1797])
+* Fix the steps for enabling the nightly flag in the pong tutorial. ([#1805])
 * Fix animation unwrap on missing animated component. ([#1773])
 
+[#1512]: https://github.com/amethyst/amethyst/issues/1512
 [#1791]: https://github.com/amethyst/amethyst/pull/1791
 [#1766]: https://github.com/amethyst/amethyst/pull/1766
 [#1719]: https://github.com/amethyst/amethyst/pull/1719
@@ -49,7 +56,10 @@ and the corresponding draw functions to `DebugLines`, to draw simple shapes with
 [#1773]: https://github.com/amethyst/amethyst/pull/1773
 [#1753]: https://github.com/amethyst/amethyst/pull/1753
 [#1720]: https://github.com/amethyst/amethyst/pull/1720
+[#1772]: https://github.com/amethyst/amethyst/pull/1772
+[#1797]: https://github.com/amethyst/amethyst/pull/1797
 [#1800]: https://github.com/amethyst/amethyst/pull/1800
+[#1805]: https://github.com/amethyst/amethyst/pull/1805
 
 ## [0.11.0] - 2019-06